# microservice-account/target/classes/api.yml
openapi: 3.0.0
info:
  title: Bank Account Microservice
  description: API to manage bank accounts.
  version: 1.0.0

servers:
  - url: http://localhost:8091

paths:

  /accounts:
    get:
      tags:
        - account
      summary: Get all accounts
      operationId: getAccounts
      responses:
        200:
          description: Successful Operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountResponse'
    post:
      tags:
        - account
      summary: Create a new account
      operationId: createAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountRequest'
      responses:
        201:
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        400:
          description: Invalid input
          content:
            text/plain:
              schema:
                type: string
                example: "Error"
  /accounts/vip:
    post:
      tags:
        - account
      summary: Create a new vip account
      operationId: createVipAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VipAccountRequest'
      responses:
        201:
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        400:
          description: Invalid input
          content:
            text/plain:
              schema:
                type: string
                example: "Error"
  /accounts/mype:
    post:
      tags:
        - account
      summary: Create a new mype account
      operationId: createMypeAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VipAccountRequest'
      responses:
        201:
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        400:
          description: Invalid input
          content:
            text/plain:
              schema:
                type: string
                example: "Error"
  /accounts/{accountNumber}:
    get:
      tags:
        - account
      summary: Find account by ID
      operationId: findAccountById
      parameters:
        - in: path
          name: accountNumber
          description: ID of account to return
          required: true
          schema:
            type: string
      responses:
        200:
          description: Operation successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        400:
          description: Invalid ID supplied
        404:
          description: Account not found
    delete:
      tags:
        - account
      summary: Delete an account by ID
      operationId: DeleteAccountById
      parameters:
        - in: path
          name: accountNumber
          required: true
          schema:
            type: string
      responses:
        204:
          description: No content
        400:
          description: Invalid ID supplied
        404:
          description: Account not found
  /accounts/client/{id}:
    get:
      tags:
        - account
      summary: Find all accounts of a client
      operationId: getAccountsOfClient
      parameters:
        - in: path
          name: id
          description: ID of account to return
          required: true
          schema:
            type: array
            items:
              type: string
      responses:
        200:
          description: Operation successful
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountResponse'
        400:
          description: Invalid ID supplied
        404:
          description: Account not found

components:
  schemas:
    VipAccountRequest:
      type: object
      properties:
        clientId:
          type: string
          example: 123
        amount:
          type: number
          format: float
          example: 10.5
    AccountRequest:
      type: object
      properties:
        accountType:
          type: string
          enum: [AHORRO, CUENTA_CORRIENTE, PLAZO_FIJO]
        clientId:
          type: array
          items:
            type: string
          example: [1234]
        clientType:
          type: string
          enum: [PERSONAL, BUSINESS]
    AccountResponse:
      type: object
      properties:
        balance:
          type: number
          format: double
          example: 10.10
        numberAccount:
          type: string
          example: 123456789
        maintenanceFee:
          type: number
          format: double
          example: 0
        monthlyMovements:
          type: integer
          example: 3
        accountType:
          type: string
          enum: [AHORRO, CUENTA_CORRIENTE, PLAZO_FIJO]
        clientId:
          type: array
          items:
            type: string
          example: [1234]

tags:
  - name: account
    description: Operations about accounts